+++ src/editor/core/plugins/max_height/index.js
@@ -5,9 +5,10 @@
  *
  */
 
-// We don't use the official repo https://github.com/valeriangalliat/markdown-it-anchor/issues/39
-import markdown_it_anchor_plugin from "markdown-it-anchor"
-import {Plugin} from "prosemirror-state";
+import {Plugin, PluginKey} from "prosemirror-state";
+import {MaxHeightState} from "./state";
+
+const pluginKey = new PluginKey('max_height');
 
 const maxHeight = {
     id: 'max-height',
@@ -18,7 +19,9 @@ const maxHeight = {
 
         context.editor.on('afterInit', () => {
             if(context.options.maxHeight) {
-                context.editor.getStage().css({'max-height': context.options.maxHeight,'overflow': 'auto'});
+                context.editor.getStage().css({'max-height': context.options.maxHeight, 'overflow': 'auto'});
+                const maxHeightState = pluginKey.getState(context.editor.view.state);
+                maxHeightState.update();
             }
         });
     },
@@ -27,38 +30,20 @@ const maxHeight = {
         let scrollActive = false;
         let initialized = false;
         return [new Plugin({
+            state: {
+                init(config, state) {
+                    return new MaxHeightState(state, {context: context});
+                },
+                apply(tr, prevPluginState, newState) {
+                    return prevPluginState;
+                }
+            },
+            key: pluginKey,
             view: (view) => {
                 return {
                     update: (view, prevState) => {
-                        let stageHeight = context.editor.getStage().innerHeight();
-
-                        if(stageHeight === oldStageHeight) {
-                            return;
-                        }
-
-                        oldStageHeight = stageHeight;
-
-                        if(!scrollActive && context.editor.getStage()[0].scrollHeight > stageHeight) {
-                            if(!initialized) {
-                                context.editor.getStage().niceScroll({
-                                    cursorwidth: "7",
-                                    cursorborder: "",
-                                    cursorcolor: "#606572",
-                                    cursoropacitymax: "0.3",
-                                    nativeparentscrolling: false,
-                                    autohidemode: false,
-                                    railpadding: {top: 2, right: 3, left: 0, bottom: 2}
-                                });
-                            }
-
-                            initialized = true;
-                            scrollActive = true;
-                            context.editor.trigger('scrollActive');
-                        } else if(scrollActive) {
-                            scrollActive = false;
-                            context.editor.trigger('scrollInactive');
-
-                        }
+                        const maxHeightState = pluginKey.getState(view.state);
+                        maxHeightState.update();
                     },
                     destroy: () => {}
                 }